remove utilization app
