+++ tock/tock/templates/base.html
@@ -23,7 +23,7 @@
 
     <script src="https://cdnjs.cloudflare.com/ajax/libs/underscore.js/1.7.0/underscore-min.js" ></script>
 
-    
+
     <script>
   (function(i,s,o,g,r,a,m){i['GoogleAnalyticsObject']=r;i[r]=i[r]||function(){
 (i[r].q=i[r].q||[]).push(arguments)},i[r].l=1*new Date();a=s.createElement(o),
@@ -35,8 +35,8 @@ m=s.getElementsByTagName(o)[0];a.async=1;a.src=g;m.parentNode.insertBefore(a,m)
   ga('set', 'forceSSL', true);
   ga('send', 'pageview');
     </script>
-    
-    
+
+
   </head>
   <body>
   <div class="container">
@@ -59,12 +59,6 @@ m=s.getElementsByTagName(o)[0];a.async=1;a.src=g;m.parentNode.insertBefore(a,m)
         <li><a href="{% url 'reportingperiod:ReportingPeriodCreateView' %}">Add reporting period</a></li>
         <li><a href="/admin">Admin panel</a>
         <li><a href="{% url 'reports:BulkTimecardList' %}">Bulk timecard CSV</a>
-        {% endif %}
-        {% if request.user.is_staff %}
-        <br />
-        <br />
-        <li><b>Supervisor:</b></li>
-        <li><a href="{% url 'utilization:GroupUtilizationView' %}">Utilization reports</a></li>
         {% endif %}
         <hr>
       </ul>
